MLABecLaplacian: Tweak kernel fusing
Always fuse the kernels in MLABecLaplacian instead of relying on the
heuristics in FabArrayBase::isFusingCandidate. This improves the performance
on GPUs because it reduces the register pressure. Because these kernels are
called many times during MLMG::solve, the one-time cost of copying the
metadata to GPU device memory is amortized.