Fix Fortran interface for FillPatch for face variables
The stride was wrong when copying MultiFab pointers from a Fortran array to
a Vector of Array of MultiFab pointers.

This also fixes cases when the user passes t_new followed by t_old, instead
of the other order. The issue was then teps was negative. This was a minor
issue, because the FillPatch function in C++ does not care about the order.