Fix Fortran interface for FillPatch for face variables (#3541)
## Summary

The stride was wrong when copying MultiFab pointers from a Fortran array
to a Vector of Array of MultiFab pointers.

This also fixes cases when the user passes t_new followed by t_old,
instead of the other order. The issue was then teps was negative. This
was a minor issue, because the FillPatch function in C++ does not care
about the order.

## Additional background

https://github.com/AMReX-Codes/amrex/issues/327#issuecomment-1716356603

## Checklist

The proposed changes:
- [x] fix a bug or incorrect behavior in AMReX
- [ ] add new capabilities to AMReX
- [ ] changes answers in the test suite to more than roundoff level
- [ ] are likely to significantly affect the results of downstream AMReX
users
- [ ] include documentation in the code and/or rst files, if appropriate