EB Data outside domain
We do not guarantee there is valid EB information outside the domain, unless
the user specifies a large enough number of ghost cells when building
EB. However, this makes things difficult for application codes with AMR,
because the number of ghost cells needed depends on the number of AMR
levels. In this commit, we try to fill some basic EB information
(EBCellFlag, volume fraction and area fraction) for all cells outside the
domain.