MLNodeABecLaplacian
Add a new Linear Opearator class MLNodeABecLaplacian.

It solves `(alpha * a - beta * (del dot b grad)) phi = rhs`, where a, phi
and rhs are nodal MultiFabs and b is cell-centered.

It works in 2D and 3D, CPU and GPU. So far it works only for single level.

There is an example of using it at Tests/LinearSolvers/ABecLaplacian_C as
prob_type=4. The test problem converges at the second order.