Free Disk Space in Github CIs (#3574)
Add a new script that removes some packages we do not need. It can free
up about 7 GB of space on Ubuntu.

Use it in nvhpc and sycl/amd CIs. These CIs often run out of space. We
don't want to use it in every CI, because it takes a few minutes to run
the script.