Adding bittree interface to improve regridding performance in octree mode (#3555)
## Summary

This PR introduces dependency on [Bittree
library](https://github.com/Flash-X/Bittree) to improve regridding
performance in octree mode.

## Additional background

Testing and development of this feature is done in sync with Flash-X and
is recorded in this [reproducibility
capsule](https://github.com/Lab-Notebooks/AMReX-Bittree-Performance).
This PR is primarily created to consolidate development work and avoid
creating multiple branches. At present using AMReX+Bittree improves
regridding performance [by a factor of
2](https://github.com/Lab-Notebooks/AMReX-Bittree-Performance/blob/14faa2212c4e5dba7fd99a6526c6937414f9c109/analysis/Performance.ipynb)
at > 20000 ranks. We hope to improve performance further using a
bittree-based distribution mapping and therefore adding a new function
`AmrMesh::MakeDistributionMap` as a place-holder.

Continuation of #2893 and #3547

## Checklist

The proposed changes:
- [ ] fix a bug or incorrect behavior in AMReX
- [x] add new capabilities to AMReX
- [ ] changes answers in the test suite to more than roundoff level
- [ ] are likely to significantly affect the results of downstream AMReX
users
- [ ] include documentation in the code and/or rst files, if appropriate

---------

Co-authored-by: Weiqun Zhang <WeiqunZhang@lbl.gov>
Co-authored-by: Tom Klosterman <https://github.com/tklos96>