Fix a memory "leak" in VisMF's persistent streams
VisMF uses a std::map to keep track of the streams it opens. Each
PersistentIFStream object also has a vector used as IO buffer. When a stream
is closed, we need to free to memory used by the buffer. The issue was it
used std::vector::clear(), which only changes size() without reducing
capacity() at all. So the memory was never freed. The issue is now fixed by
std::vector::swap.