solve_bicgstab: cut use of `s` (#3629)
## Summary

The MF named `s` seems to be an unnecessary usage of memory as the
residue `r` can fulfill its roles in the algorithm. This PR replaces `s`
with `r` and `LinComb` with `Saxpy` as appropriate.

## Additional background

This PR is part of a larger request to improve `solve_bicgstab` and
`solve_cg`.