solve_bicgstab: use fewer MFs (#3635)
## Summary

This PR cuts the number of MFs used in `solve_bicgstab`, saving on
memory and LocalCopy operations. In particular, the MFs `ph` and `sh`
are removed.

## Additional background

This is a follow up to avoid-use-of-s and other PRs to improve
`solve_bicgstab`. My own testing has shown that this PR gives the same
results as before, but regression testing should be done to verify this
in all cases.