Fix Advection_AmrCore test
The particle container should not be a static member, because it's a member
of a class derived from AmrCore, not AmrLevel. The issue of its being a
static member is that it is not deleted properly (unless we explicitly do
it).

Note that in the AmrLevel case we must make it a static member because all
AmrLevel objects need to share the same particle container.