BaseFab::lockAdd: Faster version of BaseFab::atomicAdd for OpenMP
For WarpX's Gordon Bell runs on Fugaku, Stephan Jaure of ATOS optimized
atomicAdd using pthread spin locks. This commit implements Stephan's
approach using OpenMP.

In AMReX::Initialize, we create a number of OMP locks. When BaseFab::lockAdd
is called, we loop over planes in z-direction and try to acquire a lock with
omp_test_lock. If it's successful, we can access the data in that z-plane
without worrying about race conditions. This allows us to use simd
instructions without using omp atomic add. If it's not successful, we will
try a different z-plane. The process stops till all planes are processed.