Add a linspace-like function in AMReX_Algorithm.H (#3698)
## Summary

I would like to propose to add a linspace-like function in
AMReX_Algorithm.H , since filling a container with equally spaced
numbers is a task that occurs rather frequently. This PR proposes a
possible implementation.
The container is modified in place only if it has at least 2 elements. 

## Additional background

## Checklist

The proposed changes:
- [ ] fix a bug or incorrect behavior in AMReX
- [X] add new capabilities to AMReX
- [ ] changes answers in the test suite to more than roundoff level
- [ ] are likely to significantly affect the results of downstream AMReX
users
- [ ] include documentation in the code and/or rst files, if appropriate