lockAdd: case of 2D plane in 3D
In HiPACE++, atomicAdd is used on 2d x & y planes even though AMREX_SPACEDIM
is 3. In that case, we would have all threads competing for a single lock in
the previous implementation of lockAdd. This PR fixes this use case by
having locks associated with the y-direction when the number of cells in the
z-direction is 1.