lockAdd: case of 2D plane in 3D (#3700)
## Summary

In HiPACE++, atomicAdd is used on 2d x & y planes even though
AMREX_SPACEDIM is 3. In that case, we would have all threads competing
for a single lock in the previous implementation of lockAdd. This PR
fixes this use case by having locks associated with the y-direction when
the number of cells in the z-direction is 1.

## Additional background

https://github.com/Hi-PACE/hipace/pull/1059

## Checklist

The proposed changes:
- [x] fix a bug or incorrect behavior in AMReX
- [ ] add new capabilities to AMReX
- [ ] changes answers in the test suite to more than roundoff level
- [ ] are likely to significantly affect the results of downstream AMReX
users
- [ ] include documentation in the code and/or rst files, if appropriate