Use long integer in GPU kernels (#3742)
In the current implementation of ParallelFor, we use `int` for the
linear cell index in the flatten one-dimension view. This limits the
size of the box to less than 2^30 cells (half of INT_MAX). The factor of
half comes from the grid-stride loop. This limitation has not been a
serious issue until recently because GPUs did not have that much memory.
However, the total memory on the latest GPUs has increased quite a lot.
Issues have been reported by users.

In this PR, we have started using `std::uint64_t` as the linear cell
index. An issue of using `std::uint64_t` is that 64-bit integer division
is very expensive. Fortunately, we are able to "steal" the fast division
code from https://github.com/NVIDIA/cutlass. Streaming tests have shown
very good performance. On A100, the FArrayBox version of the triad test
achieves 1.76 TB/s, which is the same as the rate of the much simpler 1D
vector version. In fact, it's slightly faster than the rate of 1.72 TB/s
from the current version in the development branch.

We have not made all kernel launches in AMReX safe for large sizes.
Nevertheless, this PR is the first step, and it will be followed up by
more PRs.