New Linear Solver: Curl of Curl (#3682)
Add a new linear solver for curl (alpha curl E) + beta E = rhs in 2D & 3D,
where E is an array of 3 MultiFabs on edges, and alpha and beta are
scalars. It supports periodic, homogeneous Dirichlet, and symmetry
boundaries. At the symmetry boundary, the normal component changes
the sign, whereas the transverse components do not.

---------

Co-authored-by: Axel Huebl <axel.huebl@plasma.ninja>