Add a logspace-like function in AMReX_Algorithm.H (#3754)
## Summary

Similarly to what was done in
https://github.com/AMReX-Codes/amrex/pull/3698, I would like to propose
to add a `logspace-like` function in `AMReX_Algorithm.H`. Indeed,
filling a container with logarithmically spaced numbers occurs rather
frequently. This PR proposes a possible implementation. As for
`linspace`, the container is modified (in place) only if it has at least
2 elements.

## Additional background

## Checklist

The proposed changes:
- [ ] fix a bug or incorrect behavior in AMReX
- [X] add new capabilities to AMReX
- [ ] changes answers in the test suite to more than roundoff level
- [ ] are likely to significantly affect the results of downstream AMReX
users
- [ ] include documentation in the code and/or rst files, if appropriate