Introduce: `AMReX_ADDRLINES` (Default: ON)
This is addign the new CMake option `AMReX_ADDRLINES` and CMake public
interface target `AMReX::Flags_ADDRLINES`. This option adds
*minimal* debug info (`-g1` / `-gline-tables-only`) to executables,
which generates more usable backtraces on crashes.

In symmetry to GNUmake, we turn this now ON. This is a breaking change.

Note that this flag still creates significant binary size overheads,
so package managers might decide to turn if off in deployments. Also,
if the increased binary sizes lead to significant startup overhead
at scale on HPC systems, we might need to reconsider the default
(for CMake and GNUmake) in the future.