Update GMRES/MLMG interface (#3779)
For the curl curl test, if beta = 1e-9 alpha/dx^2, the multigrid solver
is able to reduce the residual by 10 orders of magnitude in 10 v-cycles.
But for beta = 1e-14 alpha/dx^2, the multigrid solver's residual will
stall at about 5e-6 of the original residual. However, it can be solved
using GMRES with multigrid as preconditioner.