New constructor for MPMD::Copier (#3806)
## Summary
A new constructor, `Copier::Copier()`, is created in `MPMD::Copier`.
This constructor is ONLY for a particular setup where the utilized
`BoxArray` is exactly the same in both MPMD codes. The empty constructor
builds its `BoxArray` based on the `BoxArray` supplied by the other
code. Hence, the empty constructor MUST be used in conjunction with the
other code's constructor being `Copier::Copier(ba, dm, true)`.

---------

Co-authored-by: Bhargav Siddani <bsiddani@gigan.lbl.gov>
Co-authored-by: Weiqun Zhang <WeiqunZhang@lbl.gov>