Fix assertions involving IntVect (#3919)
A MultiFab used to have the same number of ghost cells in all
directions. This has no longer been the case for many years. But many
assertions have not been updated for this change. This commit fixes
various assertions to make it more appropriate.

Also added are overloaded functions for element-wise comparison between
an IntVect and an integer. This makes the code cleaner.