Update ParaView docs (#3934)
This PR adds to the Paraview docs describing how to create a time series
(`.series`) file to load and re-load plotfiles similar to the `.visit`
file in VisIt. A bash script to generate the `.series` file is provided
and the procedure to generate and (re)load the `.series` file is
described.