GNU Make: try nvidia-smi for CUDA arch if deviceQuery fails (#3997)
I recently updated my local workstation, and found that CUDA compilation
wasn't working properly due to `CUDA_ARCH` not being detected properly.
It turns out the `deviceQuery` utility is no longer distributed with the
toolkit since CUDA 12 or so. `nvidia-smi` can get that info, so we can
fall back to that if deviceQuery isn't present.