New parameter amrex::init_snan (#4030)
Uninitialized FArrayBox, BaseFab<double|float>,
PODVectors<double|float>, Gpu::DeviceVector<double|float>, etc. will be
initialized to signaling NaNs, if this parameter is true. The default
value for the new parameter is false unless AMREX_DEBUG or AMREX_TESTING
is true.

Note that previously this was only performed for FArrayBox, and it was
controlled by fab.init_snan.