ParmParse: Math expression support
Support math expression for integers and floating point numbers in
ParmParse. For example,

    n_cell = 128
    amrex.n_cell = {n_cell*2} 8 {16**2}

becomes

    n_cell = 128
    amrex.n_cell = 256 8 256

Note that the user must use `{}` to indicate that the expression needs to be
evaluate with amrex::Parser. If the expression contains another variable, it
will be looked up by ParmParse.

ParmParse's constructor now accepts an optional second argument,
`parser_prefix`. When a variable in a math expression is being looked up, it
will first try to find it by using the exact name of the variable. If this
attempt fails and the `ParmParse` object has a non-empty `parser_prefix`, it
will try again, this time looking up the variable by prefixing its name with
`parser_prefix` followed by a `.`. For example,

    physical_constants.c = 3.e8
    amrex.pi = 3.14
    amrex.foo = { sin(amrex.pi/2) * c**2 }

will give `foo = 8.999997146e+16` if we do

    amrex::ParmParse pp("amrex", "physical_constants");
    double foo;
    pp.get("foo", foo);

We also have to make a change to amrex::Parser and iParser. The variable
names can contain `.` now.