Add overload for `getParticleCell` that returns local cell index (#4081)
## Summary
In WarpX we have a few routines that requires the local cell index of
particles (i.e. the cell number in the current box). The AMReX function
`getParticleCell` currently only has logic to return the global cell
index (relative to the domain low end). This PR adds a new overload for
that function which does not shift the cell index based on the global
box index.

## Additional background
See https://github.com/ECP-WarpX/WarpX/pull/5118 for WarpX context.

## Checklist

The proposed changes:
- [ ] fix a bug or incorrect behavior in AMReX
- [x] add new capabilities to AMReX
- [ ] changes answers in the test suite to more than roundoff level
- [ ] are likely to significantly affect the results of downstream AMReX
users
- [ ] include documentation in the code and/or rst files, if appropriate