TinyProfiler:: Add more runtime parameters (#4100)
Add the following ParmParse parameters to TinyProfiler:

  * tiny_profiler.enabled

This parameter can be used to disable tiny profiling including CArena
memory profiling at run time.

  * tiny_profiler.memprof_enabled

This parameter can be used to disable tiny profiler's memory arena
profiling at run time. If tiny_profiler.enabled is false, this parameter
has no effects.

  * tiny_profiler.output_file

If this parameter is empty, the output of tiny profiling is dumped on
the default out stream of AMReX. If it's not empty, it specifies the
file name for the output. Note that `/dev/null` is a special name that
mean a null file.