ParmParse::prettyPrintTable
Add a new function that prints the ParmParse table without duplicates. Note
that these are not duplicates, because they have different values.

    foo.bar = 1
    foo.bar = 2

These are also not considered duplicates, even though the final results are
the same.

    c = 3
    foo.bar = c
    foo.bar = 3

This should help WarpX to reduce duplicates when printing ParmParse
parameters. However, there will still be "duplicates" like below because
they are different strings.

    geometry.prob_hi = 20.e-6 20.e-6
    geometry.prob_hi = 2.0000000000000002e-05 2.0000000000000002e-05