EB: don't abort for no-op case in unsupported addFineLevels functions (#4123)
## Summary

The `addFineLevels` function is not supported for EB2 for chk_file and
stl geometries. However, it may be called in some for some trivial cases
where it is adding 0 levels, in which case it is a no-op. There is no
reason to abort in those cases.

## Additional background

For PeleC, a work-around was put in to not call the function in the
trivial cases (https://github.com/AMReX-Combustion/PeleC/pull/771). I
was thinking about adding the same work around to address the same thing
in PeleLMeX (https://github.com/AMReX-Combustion/PeleLMeX/issues/407),
but maybe it would be better to simply allow the function to be called
in trivial cases. If there's a reason not to do this, I'll just put the
workaround in for PeleLMeX.