`ParmParse`: Prefix to `FILE`
For CI/CD workflows and out-of-source tests we often want to include
dependent inputs files via `FILE = <filename>`. For development, we
often want to run in temporary run directories but want to avoid
having to copy over the latest inputs file from a source directory.

Now, the environment variable `AMREX_INPUTS_FILE_PREFIX`
can be set to prefix every `FILE = <filename>` with a custom path.
We will use this in the CTests integration of WarpX.