MLMG: Minimum domain width (#4129)
By default, the minimum domain width at the coarsest multigrid level is
2 and 4, for non-EB and EB runs, respectively. Previously, it was set to
4 for runs compiled with EB support but without EB at run time. So the
results of those runs would be different from runs not compiled with EB
support. This is not a correctness issue. Nevertheless, for the sake of
consistence, we set the minimum domain width to 4 only if there is
non-trivial EB at run time.