AnyCTO with arbitrary number of functions (#4135)
This PR extends AnyCTO to support an arbitrary number of optimized
functions. Some constructs such as PrefixSum need multiple functions
that may need to be optimized. It is also possible to provide no GPU
function, in which case the compile time parameters will be directly
given to the CPU function.