New function for domain decomposition
Add a new function that decomposition a Box into BoxArray. The number of
Boxes in the returned BoxArray matches the given nboxes argument, unless the
domain is too small. We aim to decompose the domain into subdomains that are
as cubic as possible, even if this results in Boxes with odd numbers of
cells. Thus, this function is generally not suited for applications with
multiple AMR levels or for multigrid solvers. However, it could be useful
for single-level applications that prefer exactly one Box per process.