Bounded sampling (#4195)
The `get_line_data` and `get_slice_data` routines do not currently allow
for a natural way to bound the returned MultiFab. In the former case,
the line length is dictated by the bottom and top of the boxarray at
that level. In the latter case, the plane is dictated by the problem
domain at that level.

This PR adds an optional `Box` argument to the `get_line_data` call and
an optional `RealBox` argument to the `get_slice_data` call. These
optional arguments bound the returned MuliFab. This feature has
application in the ERF code in [PR
1897](https://github.com/erf-model/ERF/pull/1897) where it is of great
value to sample velocities in planes/lines just upstream of a wind
turbine. Since the wind turbine is small in relation to the problem
domain, the bounding feature ensures the returned data is actually what
the user wants.