PC Names: Same Type, Multiple Vars/Instances
If the same particle container type is used to create
multiple particle species, then the compile-time defined
names were only added for the first particle species (instance)
of the type and then skipped for any later species (another
instance of the same type).

This is due to the global variable (so many globals...) used in
the type init. This fixes the problem by moving the instance
related name logic out of the once-per-type logic `if` branch.