GPU Mapping
For perlmutter and frontier, if there are multiple device available, we will
try to map GPUs to the closest core.

For an FFT test on perlmutter using 256 nodes, the correct mapping reduced
the run time from 0.172 to 0.127. Note that you can achieve the similar
effect with `srun ... bash -c "export
CUDA_VISIBLE_DEVICES=\$((3-SLURM_LOCALID)); ..."` by manually limiting the
number of visible devices. But in this commit, we are trying to do this
automatically for the user. Also note that MPI appears to crash with
gpu-bind=closest on perlmutter. So we need to use gpu-bind=none.

For frontier, you could use gpu-bind=closest. But if your use gpu-bind=none,
this commit will try to do the correct mapping for you.

In this commit, we also removed the old machin stuff and added new code for
machine detection.