update docs for HYPRE installation (#4372)
AMReX does not currently support the latest versions of hypre available
from github, so instruct users to checkout the latest release that is
supported (v2.32.0).

Also correct some formatting issues with the critical `--with-gpu-arch`
flag and provide a debugging hint for what happens when built for the
wrong architecture.

Add a recommendation for building with the `--enable-unified-memory`
memory flag