ParmParse: Type Hints (#4364)
## Summary

Today, we use the `ParmParse` object not only anymore to read text-based
key-value pairs from input files and CLI options, but additionally and
sometimes exclusively populate it heavily from the C++ and Python APIs
directly.

This proposes a concept to store the type of the latest `add` with each
entry, which can help to expose `ParmParse` options, e.g., when
exporting them to Python dictionaries and other file formats (YAML,
TOML, JSON, XML, etc.). When a typehint is found, we can cast to the
correct type automatically for the value, instead of returning a string.

## Additional background

https://github.com/AMReX-Codes/pyamrex/pull/417#issuecomment-2699220142

## Checklist

The proposed changes:
- [ ] fix a bug or incorrect behavior in AMReX
- [x] add new capabilities to AMReX
- [ ] changes answers in the test suite to more than roundoff level
- [ ] are likely to significantly affect the results of downstream AMReX
users
- [ ] include documentation in the code and/or rst files, if appropriate