ReduceToPlaneMF
Add a new Reduce to plane function that returns the results in a pair of
MultiFabs. The first MultiFab is distributed the same as the original data
MultiFab with each Box flatten to 1 cell. The Fabs of the first MultiFab
contain box-local reduction results. The second MultiFab has only a single
cell in the reduction direction, and it contains global reduction results
distributed in the other two directions.