fix multi-level ParticleToMesh when zero_out_input is false (#4396)
## Summary
When more than one level is passed to `amrex::ParticleToMesh`, the input
MultiFabs are always zeroed out before they are overwritten with the
particle deposition, even if `zero_out_input` is set to false. This
fixes the behavior for the multi-level case when `zero_out_input` is
false.

## Additional background

## Checklist

The proposed changes:
- [x] fix a bug or incorrect behavior in AMReX
- [ ] add new capabilities to AMReX
- [ ] changes answers in the test suite to more than roundoff level
- [ ] are likely to significantly affect the results of downstream AMReX
users
- [ ] include documentation in the code and/or rst files, if appropriate