Combine back-to-back saxpy operations, as well as saxpy next to xpay (#4366)
This is intended to reduce the number of kernels created on GPUs. The
calls to these new kernels are only in MLCGSolver because that is what
we have been looking at, but they likely could be added to other places
in the code. This came out of the Open Hackathon two weeks ago. @marchdf
can provide more detail about the performance improvement this was shown
to provide.

---------

Co-authored-by: Weiqun Zhang <weiqunzhang@lbl.gov>