Add utility to create non-periodic turbulence files from planes of ou… (#4418)
Add a utility to assemble FAB files (appended with the tilmestep of the
plane) into "non-periodic turbulence inflow" files. A similar capability
exists to create a turbinflow file from a snapshot of a triply periodic
dataset. Rather than assuming a Taylor hypothesis and inflowing into a
domain at a fixed advection speed, this version provides a mechanism to
essentially couple the outflow of one run as inflow into another, thus
removing the assumption of a uniformly moving reference frame.

## Summary
Given a set of FAB files, each of which is appended with a timestamp,
assemble a single folder/file to be used as source for inflowing a
velocity field into an AMReX calculation. The FAB data must be from a
uniform grid (usually taken from level 0 data of a precursor run), and
the timestamps are assumed to be increasing in the file list given here
as input. Example code to read and use files of this format can be found
[here](https://github.com/AMReX-Combustion/PelePhysics/tree/development/Source/Utility/TurbInflow)
for example.

## Additional background
This (re-)implements the (ambiguously named) "swirl-type" turbulence
that was dropped years prior when the turbulence inlet capability was
upgraded for GPU implementation, adding the file generation tool in the
same folder as the existing tool that creates data for the periodic
(Taylor hypothesis) version.

## Checklist

The proposed changes:
- [ ] fix a bug or incorrect behavior in AMReX
- [x] add new capabilities to AMReX
- [ ] changes answers in the test suite to more than roundoff level
- [ ] are likely to significantly affect the results of downstream AMReX
users
- [ ] include documentation in the code and/or rst files, if appropriate