Fix safety of amrex::launch (#4419)
It's known that some versions of amrex::launch functions are not safe
because they use AMREX_GPU_MAX_THREADS as the launch bound. Although
they have an assertion for the number of threads per block, the unsafe
amrex::launch functions are used in a few places resulting in runtime
failures. This PR fixes these unsafe launch functions by removing the
launch bound when we do not know the number of threads per block at
compile time. We also pass the number of threads per block as a template
parameter so that it can be used as the launch bound, when it's
possible.