CArena free unused memory before allocate (#4451)
## Summary

In specific special cases, an application may ask for chunks of memory
with slightly increasing size. If these do not fit into the existing
space of the arena, a new chunk of system memory has to be allocated
every time. The old chunks that are slightly too small to be reused
could accumulate and consume a lot of memory and would never get freed
with the default release threshold of max long. This PR adds a check to
free unused system memory before allocating new memory, while still
keeping the total amount of memory the same to avoid excessive system
allocations and frees.

## Additional background

The fix from #4044 is reimplemented differently in the PR so that
freeUnused_protected can be called.