ParmParse: Fix backward compatibility issue in #4457 (#4460)
We made some changes in reading IntVect and Box in #4457 to support
inline comments. However, WarpX uses unquoted strings for parser
expressions. These strings often contains parentheses. We need to
preserve the white spaces inside parentheses, otherwise `(z and z)` will
become `(zandz)`.

Note that this is not an issue for quoted strings.