Add new amrex::Initialize (#4476)
In #4463, MPI data types were removed from the global namespace in
non-MPI builds. This broke several codes. Some codes use
`MPI_COMM_WORLD` when calling amrex::Initialize because that was the
recommended way of modifying the default ParmParse parameters. In this
PR, we add a new amrex::Initialize that allows the users to modify
ParmParse parameters without providing an MPI Communicator.