Workarounds for Lustre I/O issues (#4426)
This adds workarounds for Lustre I/O write issues at scale (~128 nodes
or more):
* moving the flush after FAB writes until after the whole MultiFab has
been written,
* removing the flush after particle real data and particle int data (a
flush is still done in amrex::NFilesIter), and
* adding a `ParallelDescriptor::Barrier()` after writing each level
(both for MultiFabs and particles).

This reduces plotfile write time on 1024 nodes on Frontier from 30+
minutes to 1 minute. It performs best with a stripe count of 1 and
stripe size of 16M and 1 file per node.