Delay gpu stream sync in MFIter
Previously we performed gpu stream sync during MFIter::Initialize in case
that kernels on non-default streams inside MFIter loop might depend on the
result of kernels on the default amrex stream launched before the MFIter
loop. Now we delay the sync if needed until the end of the first iteration.